requirements.txt
asgiref==3.8.1
boto3==1.35.23
botocore==1.35.23
Django==4.2.16
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==23.0.0
inflection==0.5.1
jmespath==1.0.1
packaging==24.1
psycopg2==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
s3transfer==0.10.2
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.15
